[437582] Disable broken doc contributions
The registered toc entries are defined in files which are missing from
the build, causing errors in the help system. For Oxygen, disable them
completely to avoid the warnings. In a future version they will be
re-included, but this will need some review of the actual content that
they contribute, which requires more time that I have available right
now.

Bug: 437582
Change-Id: I361cdd73ed9261daed06c095ff32b0febfaf91f8
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>